refactor: move the common pipeline methods in a separate file
